Marconi.Sidechain.Api.HttpServer

Documentation

bootstrap :: SidechainEnv -> IO ()

Bootstraps the HTTP server

marconiApp :: SidechainEnv -> Application

jsonRpcServer

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> Server JsonRpcAPI

restApiServer :: SidechainEnv -> Server RestAPI

httpRpcServer

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> Server API

echo :: String -> Handler (Either (JsonRpcErr String) String)

Echos message back as a Jsonrpc response. Used for testing the server.

getTimeHandler :: Handler String

Echos current time as REST response. Used for testing the http server outside of jsonrpc protocol.

getTargetAddressesHandler

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> Handler [Text]

Prints TargetAddresses Bech32 representation to the console

getTargetAddressesQueryHandler

Arguments

:: SidechainEnv
database configuration

-> String
Will always be an empty string as we are ignoring this param, and returning everything

-> Handler (Either (JsonRpcErr String) [Text])

prints TargetAddresses Bech32 representation as thru JsonRpc

getCurrentSyncedPointHandler

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> String
Will always be an empty string as we are ignoring this param, and returning everything

-> Handler (Either (JsonRpcErr String) CurrentSyncedPointResult)

Handler for retrieving current synced chain point.

getAddressUtxoHandler

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> TxOutAtQuery
Bech32 addressCredential and a slotNumber

-> Handler (Either (JsonRpcErr String) AddressUtxoResult)

Handler for retrieving UTXOs by Address

getMintingPolicyHashTxHandler

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> String
Minting policy hash

-> Handler (Either (JsonRpcErr String) MintingPolicyHashTxResult)

Handler for retrieving Txs by Minting Policy Hash.

getEpochStakePoolDelegationHandler

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> Word64
EpochNo

-> Handler (Either (JsonRpcErr String) EpochStakePoolDelegationResult)

Handler for retrieving stake pool delegation per epoch

getEpochNonceHandler

Arguments

:: SidechainEnv
Utxo Environment to access Utxo Storage running on the marconi thread

-> Word64
EpochNo

-> Handler (Either (JsonRpcErr String) EpochNonceResult)

Handler for retrieving stake pool delegation per epoch

toRpcErr :: QueryExceptions -> JsonRpcErr String

Convert to JSON-RPC protocol error.