Marconi.Sidechain.Api.Query.Indexers.EpochState

Documentation

initializeEnv

Arguments

:: IO EpochStateIndexerEnv
returns Query runtime environment

Bootstraps the EpochState query environment. The module is responsible for accessing SQLite for queries. The main issue we try to avoid here is mixing inserts and quries in SQLite to avoid locking the database

updateEnvState :: TMVar (State EpochStateHandle) -> State EpochStateHandle -> STM ()

querySDDByEpochNo

Arguments

:: SidechainEnv
Query run time environment

-> Word64
Bech32 Address

-> IO (Either QueryExceptions EpochStakePoolDelegationResult)
Plutus address conversion error may occur

Retrieve SDD (stakepool delegation distribution) associated at the given EpochNo. We return an empty list if the EpochNo is not found.

queryNonceByEpochNo

Arguments

:: SidechainEnv
Query run time environment

-> Word64
Bech32 Address

-> IO (Either QueryExceptions EpochNonceResult)
Plutus address conversion error may occur

Retrieve the nonce associated at the given EpochNo We return an empty list if the EpochNo is not found.