Marconi.Sidechain.Api.Query.Indexers.Utxo

Documentation

initializeEnv

Arguments

:: Maybe TargetAddresses
user provided target addresses

-> IO AddressUtxoIndexerEnv
returns Query runtime environment

Bootstraps the utxo query environment. The module is responsible for accessing SQLite for quries. The main issue we try to avoid here is mixing inserts and quries in SQLite to avoid locking the database

currentSyncedPoint

Arguments

:: AddressUtxoIndexerEnv
Query run time environment

-> IO (Either QueryExceptions CurrentSyncedPointResult)
Wrong result type are unlikely but must be handled

Retrieve the current synced point of the utxo indexer

findByAddress

Arguments

:: AddressUtxoIndexerEnv
Query run time environment

-> AddressAny
Cardano address to query

-> Maybe SlotNo
The upper slot number we want to query

-> IO (Either QueryExceptions [UtxoRow])

Query utxos by Address Address conversion error from Bech32 may occur

findByBech32Address

Arguments

:: AddressUtxoIndexerEnv
Query run time environment

-> Text
Bech32 Address

-> IO (Either QueryExceptions AddressUtxoResult)
Plutus address conversion error may occur

Retrieve Utxos associated with the given address We return an empty list if no address is not found

findByBech32AddressAtSlot

Arguments

:: AddressUtxoIndexerEnv
Query run time environment

-> Text
Bech32 Address

-> Maybe Word64
Slot number to look at

-> IO (Either QueryExceptions AddressUtxoResult)
Plutus address conversion error may occur

Retrieve Utxos associated with the given address We return an empty list if no address is not found

reportQueryAddresses :: AddressUtxoIndexerEnv -> IO [Address ShelleyAddr]

report target addresses Used by JSON-RPC

data UtxoRow

Constructors

UtxoRow

Fields

_urUtxo :: !Utxo

_urSlotNo :: !SlotNo

_urBlockHash :: !(Hash BlockHeader)

type UtxoIndexer = State UtxoHandle

reportBech32Addresses :: AddressUtxoIndexerEnv -> [Text]

withQueryAction

Arguments

:: AddressUtxoIndexerEnv
Query run time environment

-> StorableQuery UtxoHandle
Address and slot to query

-> IO (Either QueryExceptions [UtxoRow])

Execute the query function We must stop the utxo inserts before doing the query

updateEnvState :: AddressUtxoIndexerEnv -> UtxoIndexer -> STM ()