Marconi.Sidechain.Bootstrap

Description

This module bootstraps the marconi-sidechain JSON RPC server, it acts as a glue conntecting the JSON-RPC, HttpServer, marconiIndexer, and marconi cache

Documentation

initializeSidechainEnv :: Maybe Port -> Maybe TargetAddresses -> IO SidechainEnv

Bootstraps the JSON-RPC http server with appropriate settings and marconi cache this is just a wrapper for the bootstrapHttp in json-rpc package

bootstrapIndexers :: CliArgs -> SidechainEnv -> IO ()

Marconi cardano blockchain indexer